% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowed_species}
\alias{allowed_species}
\title{allowed_species}
\format{
An object of class \code{character} of length 77.
}
\usage{
allowed_species
}
\description{
A vector of character containing all species names that are allowed in Sötebasen.
}
\keyword{datasets}
