% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{calc_depth}
\alias{calc_depth}
\title{Calculate water level}
\usage{
calc_depth(P_water, P_ref, temp_water)
}
\arguments{
\item{P_water}{pressure in water in kPa}

\item{P_ref}{reference air pressure in kPa}

\item{temp_water}{water temperature in degrees Celsius}
}
\value{
A numeric. water level in meters
}
\description{
Calculate water level from barimetric pressure at meassured bottom,
}
