% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{can_coerce_numeric}
\alias{can_coerce_numeric}
\title{can_coerce_numeric}
\usage{
can_coerce_numeric(x)
}
\arguments{
\item{x}{}
}
\value{
Returns TRUE if all items can be coerced as numeric
}
\description{
Checks if a vector either is numeric or can be coerced to numeric without introducing NA
}
