% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{dec2hex}
\alias{dec2hex}
\title{Decimal to hex pittag string}
\usage{
dec2hex(tags)
}
\arguments{
\item{tags}{character vector of tags in hex format}
}
\value{
character vector of tags in decimal format
}
\description{
Convert vector of pittags in dec to hex
}
