% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event2Behandling}
\alias{event2Behandling}
\title{Translate event codes to Sötebasens strings for Behandling}
\usage{
event2Behandling(x)
}
\arguments{
\item{x}{numeric}
}
\value{
Text representation of event codes defined for Smoltreg
}
\description{
Translate event codes to Sötebasens strings for Behandling
}
\examples{
ev <- c(min(Smoltreg::event):max(Smoltreg::event))
sb_ev <- event2Behandling(ev)
sb_ev
}
