% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{fulton}
\alias{fulton}
\title{Fulton condition factor}
\usage{
fulton(w, l)
}
\arguments{
\item{w}{Weight in grams}

\item{l}{Length in mm}
}
\value{
A numeric
}
\description{
Return Fulton's condition factor
}
