% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{hex2dec}
\alias{hex2dec}
\title{Hex to dec pittag string}
\usage{
hex2dec(tags)
}
\arguments{
\item{tags}{character vector of tags in hex format}
}
\value{
character vector of tags in decimal format
}
\description{
Convert vector of pittags in hex to dec
}
