% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{impute_date_time}
\alias{impute_date_time}
\title{impute_date_time}
\usage{
impute_date_time(x)
}
\arguments{
\item{x}{Vector with dates}
}
\value{
imputed vector with dates
}
\description{
Impute a time vector with dates using "last observation carried forward"
if package imputeTS is installed
}
