% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{is.possible.numeric}
\alias{is.possible.numeric}
\title{is.possible.numeric}
\usage{
is.possible.numeric(x)
}
\arguments{
\item{x}{}
}
\value{
Returns TRUE if all items can be coerced as numeric
}
\description{
Checks if a vector either is numeric or can be coerced to numeric without introducing NA
}
