% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{mean_nooutliers}
\alias{mean_nooutliers}
\title{mean_nooutliers}
\usage{
mean_nooutliers(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Remove outliers from x using boxplot.stats and return the mean of remaining data
}
