% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_envdata}
\alias{read_envdata}
\title{Read temp and pressure data}
\usage{
read_envdata(
  xlsxfile,
  firstdate,
  lastdate,
  sheet1 = "Envlogger_water",
  sheet2 = "Envlogger_land"
)
}
\arguments{
\item{xlsxfile}{Name of excel file}

\item{firstdate}{Date for first date in returned time series}

\item{lastdate}{Date for last date in returned time series}

\item{sheet1}{Name of sheet with data from logger in water}

\item{sheet2}{Name of sheet with data from logger in air}
}
\value{

}
\description{
Read data with water temperature and pressure + reference pressure in air.
Calculate mean water depth and water temperature per day.
}
