% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fish}
\alias{read_fish}
\title{Read sheet with fish data and do some basic cleanup.}
\usage{
read_fish(
  xlsxfile,
  dummy_tags = NULL,
  sheet = "Fiskdata",
  date_formats = c("\%m-\%d-\%Y \%H:\%M:\%S", "\%Y-\%m-\%d \%H:\%M:\%S",
    "\%m/\%d/\%y.\%H:\%M:\%S")
)
}
\arguments{
\item{xlsxfile}{Name on excel-file following the "Smoltreg"-format}

\item{dummy_tags}{Vector of character with tags used as dummies that should be removed}

\item{sheet}{Name of sheet with the fish data. Default = "Fiskdata"}

\item{date_formats}{Character vector with DateTime formats that should be tried when reading dates}
}
\value{
A data frame with fishdata from the smolt trap. Basic cleanup done.
}
\description{
Read an excel-file in the \strong{Smoltreg}-format and do some basic cleanup before a data.frame
is returned.
If a vector of dummy_tags is given all tags matching are removed. The species name
is standardized to capital first letter and the rest lowercase. Columns \strong{smoltstat} and
\strong{genid} are uppercased. The column \strong{date_time} is standarised to POSIXct and
missing values are imputed using a "last observation carried forward" algorithm.
Missing Any NA in column \strong{event} are replaced with the UNKNOWN event code. Recaptured
fish without species get the species set to the same as the MARKED event for that pit tag.
}
