% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_meta}
\alias{read_meta}
\title{Read sheet with meta data}
\usage{
read_meta(xlsxfile)
}
\arguments{
\item{xlsxfile}{Name on excel-file following the "Smoltreg"-format}
}
\value{
A list with variables from sheets "Metadata" and "Metadata2"
}
\description{
Read sheet with meta data
}
