% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{remove_empty_row}
\usage{
remove_empty_rows(df)
}
\arguments{
\item{df}{}
}
\value{
a data.frame
}
\description{
Remove rows from a data.frame when all fields in a row are NA
}
