% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoltreg-package.R
\docType{package}
\name{Smoltreg-package}
\alias{Smoltreg}
\alias{Smoltreg-package}
\title{Smoltreg: Check data quality of data files delivered from smolt traps}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anders Kagervall \email{anders.kagervall} (\href{https://orcid.org/0000-0003-4790-2696}{ORCID})

}
\keyword{internal}
