% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Smoltreg_limits}
\alias{Smoltreg_limits}
\title{Set or get limits for allowed values for some values}
\usage{
Smoltreg_limits(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get limits for allowed values for some values
}
\section{Supported options}{

The following options are supported
\itemize{
\item{\code{min_k}}{(\code{numeric};0.5) Minimum Fulton condition }
\item{\code{max_k}}{(\code{numeric};1.5) Maximum Fulton condition }
\item{\code{minlength}}{(\code{numeric};1.5) Minimum length to be considered as smolt }
\item{\code{maxlength}}{(\code{numeric};1.5) Maximum length to be considered as smolt}
}
}

