% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{strip_time}
\alias{strip_time}
\title{strip_time set components minutes and seconds to zero}
\usage{
strip_time(t)
}
\arguments{
\item{t}{a vector that can be coerced to time with as.POSIXlt}
}
\value{
a vector of POSIXct
}
\description{
strip_time set components minutes and seconds to zero
}
