% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{water_density}
\alias{water_density}
\title{water_density}
\usage{
water_density(temp_water)
}
\arguments{
\item{temp_water}{}
}
\value{
water density at temperature
}
\description{
Return water density at water temp temp_water
Borrowed from https://github.com/GLEON/rLakeAnalyzer/blob/master/R/water.density.R
}
