% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{which_coerce_NA}
\alias{which_coerce_NA}
\title{which_coerce_NA}
\usage{
which_coerce_NA(x, allow.orig.NA = TRUE, orig.NA = -989898)
}
\arguments{
\item{x}{}

\item{allow.orig.NA}{Should NA be allowed. E.g. only NA introduced by as.numeric should be returned}

\item{orig.NA}{if allow.orig.NA id TRUE this number is internally used for NA in the input}
}
\value{
Return index of the elements that would coerce to NA in a vector
}
\description{
Return index of the elements that would coerce to NA in a vector
}
