% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event2Behandling}
\alias{event2Behandling}
\title{Translate event codes to Sötebasens strings for Behandling}
\usage{
event2Behandling(x)

event2Behandling(x)
}
\arguments{
\item{x}{event codes}
}
\value{
Text representation of event codes defined for Smoltreg

character strings
}
\description{
Translate event codes to Sötebasens strings for Behandling

Convert event code to Sötebasen string
}
\examples{
ev <- c(min(Smoltreg::event):max(Smoltreg::event))
sb_ev <- event2Behandling(ev)
sb_ev
event2Behandling(c(0,1,2,3,4))
}
