% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{event2Behandling}
\alias{event2Behandling}
\title{Convert event code to Sötebasen string}
\usage{
event2Behandling(x)
}
\arguments{
\item{x}{event codes}
}
\value{
character strings
}
\description{
Convert event code to Sötebasen string
}
\examples{
event2Behandling(c(0,1,2,3,4))
}
