% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fish}
\alias{read_fish}
\title{Read sheet with fish data and do some basic cleanup.}
\usage{
read_fish(
  xlsxfile,
  dummy_tags = NULL,
  sheet = "Fiskdata",
  date_formats = c("\%m-\%d-\%Y \%H:\%M:\%S", "\%Y-\%m-\%d \%H:\%M:\%S",
    "\%m/\%d/\%y.\%H:\%M:\%S")
)
}
\arguments{
\item{xlsxfile}{Name on excel-file following the "Smoltreg"-format}

\item{dummy_tags}{Vector of character with tags used as dummies that shoud be removed}

\item{sheet}{Name of sheet with the fish data. Default = "Fiskdata"}

\item{date_formats}{Character vector with DateTime formats that should be tried when reading dates}
}
\value{
A data frame with fishdata from the smolt trap. Basic cleanup done.
}
\description{
Read sheet with fish data and do some basic cleanup.
}
