% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoltreg_event_doc.R
\docType{data}
\name{Smoltreg_event}
\alias{Smoltreg_event}
\title{Smoltreg_events}
\format{
A data frame with 386469 rows and 29 variables:
\describe{
\item{serial}{serial number, numeric}
\item{tag_number}{carlin tag number, character}
}
}
\usage{
Smoltreg_event
}
\description{
Constants for all allowed numbers in column "HÄNDELSE" in the Smoltreg format
}
\keyword{datasets}
